requirements.txt
asgiref==3.9.1
crispy-bootstrap5==2025.6
diff-match-patch==20241021
Django==5.2.5
django-crispy-forms==2.4
django-formtools==2.5.1
django-import-export==4.3.9
et_xmlfile==2.0.0
mysqlclient==2.2.7
openpyxl==3.1.5
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
